api: Create an initial client abstraction
The HMRC APIs are accessed via HTTPS endpoints.  The mechanism for
constructing the request and parsing the response is generally
consistent across different endpoints, and so can viably be expressed
by defining an introspectable endpoint object as a class attribute,
and allowing that endpoint object to be called as though it were an
instance method.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>