auth: Support use of the out-of-band redirect URI
Default to the out-of-band redirect URI "urn:ietf:wg:oauth:2.0:oob" if
no redirect URI is explicitly specified.

Note that oauthlib will refuse to use non-HTTPS URIs unless the
environment variable OAUTHLIB_INSECURE_TRANSPORT is set.  We therefore
set this variable for the duration of the token fetch operation if
using the out-of-band URI.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>