auth: Implement test user authorization as an oauthlib.oauth2.Client
Test user authorization uses the username and password defined for the
test user, and so is functionally equivalent to the RFC6749 Resource
Owner Password Credentials grant type as implemented by
oauthlib.oauth2.LegacyApplicationClient.

Choose to implement the test user authorization as a subclass of
oauthlib.oauth2.WebApplicationClient that can be used in the same way
as oauthlib.oauth2.LegacyApplicationClient.  This allows for the
possibility that HMRC will in future allow the use of the Resource
Owner Password Credentials grant type for test user authorization, at
which point the TestUserAuthClient class can be deleted.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>