api: Ignore path parameters with an explicit value of None
Allow for the existence of code such as

  def get_obligations(vrn=None):
    return self.vat.obligations(vrn=vrn)

where the intention is that an explicit value of None means to behave
as though the argument were not specified and so fall back to using
the attribute as defined on the client object (i.e. self.vat.vrn in
the above example).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>