test: Treat a missing server token as None rather than empty string
The HmrcSession constructor expects a missing server token to be
passed as None (or omitted entirely).  Erroneously passing an empty
string will cause a failed request when attempting to issue a request
to an open access endpoint.

Fix by ensuring that an empty SERVER_TOKEN test environment variable
is treated as None rather than as an empty string.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>