api: Use HMRC-issued client ID and "secret" values by default
This application has finally been authorised by the HMRC Developer
Hub.  Default to using the issued client ID and client "secret", if
not explicitly overridden.

Note that the client "secret" is associated with the codebase (rather
than with any individual user).  It does not itself grant access to
any user data: it merely allows the codebase to access the
authentication endpoints and to therefore direct the user through the
real authentication process.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>