test: Avoid spurious pylint arguments-differ warnings
Newer versions of pylint seem to spuriously complain about the number
of arguments differing when a static method is overridden by a
non-static method (e.g. because the base class method is marked as a
static method to avoid a no-self-use warning).

Remove the static method decorators and disable the no-self-use
warning to avoid spurious pylint complaints.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>